backend/requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.22
Flask-Migrate==3.0.1
stripe==2.56.0
twilio==6.57.0
pyyaml==5.4.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
flask-cors==3.0.10
passlib==1.7.4
Flask-JWT-Extended==4.3.1
bcrypt==3.2.0
requests==2.26.0
PyJWT==2.1.0

